New Overlay Layer and Fixed Odd AimOffset Wrapping
An Additional Overlay Layering Layer was introduced for Adding motion over the entirety of the Upperbody in a Synchronized Manner, this excludes the hands for accuracy which is fine as long as you keep it's weight low.

Odd AimOffset Wrapping where if he is looking down he will interpolate so that he is looking up half-way through a wrap-around to the other side. We want him to stay looking down while he moves over to the other side, this is now fixed.

Linked Layer Children: Linked Layers are now setup as child classes of the Pistol/Rifle states allowing you to easily swap out the Poses themselves and make multiple Children for different poses while still using the same Logic to drive them.

Other Adjustments, I moved Layering Data over into the Linked Anim Layer Folder since they are specific To That