Fixed Snapping Issue After Traversing
DefaultSlot Assumes it's placed before OffsetRootBone so that it can Accumulate the Rotation of the Montage as it's Blending out or else the Rotation will be Released upon finishing the Montage.

To Get Around Conflicts with Layering I simply decided to place the player in Unequipped Mode before Traversing and then setting him back after it's done.